Split up graphic
