+++ src/Graphic/Graphic.tsx
@@ -1,35 +1,21 @@
-import React, { Component } from "react";
+import { Component } from "react";
 import "./Graphic.css";
+import { Particle, animate, createParticle } from "./utils";
 
-interface IParticle {
-  x: number;
-  y: number;
-  velocityX: number;
-  velocityY: number;
-  rgba: string;
-  size: number;
-}
-
-interface IGraphicState {
-  particles: IParticle[];
+type GraphicState = {
+  particles: Particle[];
   numParticles: number;
-  maxVelocity: number;
-  maxSize: number;
-  minSize: number;
   width: number;
   height: number;
 }
 
-export class Graphic extends Component <{}, IGraphicState> {
+export class Graphic extends Component <{}, GraphicState> {
 
   constructor(props: any) {
     super(props);
 
     this.state = {
       height: 1080,
-      maxSize: 2,
-      maxVelocity: 3,
-      minSize: 0.3,
       numParticles: 20,
       particles: [],
       width: 1920,
@@ -45,28 +31,16 @@ export class Graphic extends Component <{}, IGraphicState> {
 
   public componentDidMount() {
 
-    const particles: IParticle[] = [];
+    const particles: Particle[] = [];
 
     const {
       height,
-      maxSize,
-      minSize,
-      maxVelocity,
       numParticles,
       width,
     } = this.state;
 
     for (let i = 1; i <= numParticles; i++) {
-
-      const pt: IParticle = {
-        rgba: randomColour(),
-        size: randomNumber(minSize, maxSize),
-        velocityX: randomNumber(-maxVelocity, maxVelocity),
-        velocityY: randomNumber(-maxVelocity, maxVelocity),
-        x: width / 2,
-        y: height / 2,
-      };
-
+      const pt = createParticle({ width, height }); 
       particles.push(pt);
     }
 
@@ -107,39 +81,7 @@ export class Graphic extends Component <{}, IGraphicState> {
       height,
     } = this.state;
 
-    context.clearRect(0, 0, width, height);
-    context.save();
-
-    // update the canvas size
-    // https://developer.mozilla.org/en-US/docs/Web/API/Window/devicePixelRatio
-    context.canvas.style.width = width + "px";
-    context.canvas.style.height = height + "px";
-
-    const scale = window.devicePixelRatio;
-    context.canvas.width = width * scale;
-    context.canvas.height = height * scale;
-
-    context.scale(scale, scale);
-
-    this.state.particles.forEach((particle) => {
-
-      drawParticle(particle, context);
-
-      particle.x += particle.velocityX;
-      particle.y += particle.velocityY;
-
-      if (particle.x < 0 || particle.x > width) {
-        particle.velocityX = -particle.velocityX;
-        particle.x = particle.x < 0 ? 1 : width;
-      }
-
-      if (particle.y < 0 || particle.y > height) {
-        particle.velocityY = -particle.velocityY;
-        particle.y = particle.y < 0 ? 1 : height;
-      }
-
-      return particle;
-    });
+    animate(context, width, height, this.state.particles);
   }
 
   public render() {
@@ -147,26 +89,4 @@ export class Graphic extends Component <{}, IGraphicState> {
       <canvas className="particle" ref="canvas"/>
     );
   }
-}
-
-function randomNumber(min: number, max: number): number {
-  return (Math.random() * (max - min)) + min;
-}
-
-function randomColour(): string {
-  const rand = randomNumber(0, 10);
-
-  return rand < 5 ? "#282c34" : "#B9009A";
-}
-
-function drawParticle(pt: IParticle, context: any) {
-  const size = pt.size;
-
-  context.beginPath();
-  context.arc(pt.x, pt.y, size, 0, 2 * Math.PI);
-  context.strokeStyle = pt.rgba;
-  context.stroke();
-
-  context.fillStyle = pt.rgba;
-  context.fill();
 }
